Check integrity of webrtc download in build script (#6)
We see periodic failures of zed-industries/zed CI when the `cc` builder
in the `webrtc-sys` build script fails to find some webrtc header files
that are supposed to be downloaded using this function:


https://github.com/zed-industries/livekit-rust-sdks/blob/d2eade7a6b15d6dbdb38ba12a1ff7bf07fcebba4/webrtc-sys/build/src/lib.rs#L179

This downloads a zip from the assets attached to a specific
livekit/rust-sdks release, and unpacks it under OUT_DIR. My guess is
that sometimes the download step fails and produces a corrupt downloaded
zip, without signaling the failure in a way that would stop the build.
This PR tries to fix that by having the download step check the SHA-256
digest of the downloaded zip (using a manually-computed list of good
digests; I'd like to use digests from the GitHub API but the particular
artifacts we download don't have them).

If we confirm that this is the source of the problem, I'll follow up
with a change to retry the download in the build script.