Remove remnants of webrtc_dir caching (#7)
Ensure that we download the webrtc release anew each time
download_webrtc is called (from either the livekit-ffi build script or
the webrtc-sys build script). This is another stab at fixing the build
errors we've been seeing in Zed CI, or at least making them easier to
understand.

Also reverts #6 since it seems the integrity of the downloaded zip was a
red herring.